Hey Guys.  This is my first project.  It is the beginning of a Settlers of Catan game, made for the Python terminal.  Eventually I would like to have the fully functional game (as well as a GUI).
Some of the missing components of the real game:
Ports: both in the graphics and functionality.
Cards: their defintion, their purchase, and their use.

Some of the things I would like to add:
Private message functionality
Public message functionality
Retrieveable statistics

Most of all, there needs to be a way to distribute the needed text to each player on a different computer.  I would like it to work over internet, but if it only works local, I would be content.

There are numerous bugs, especially with the "graphics" and I will try to get around to them asap (but it is really boring programming, so don't hold your breath).

There are some really clunky if then trees that I would like to get rid of, but am too lazy to figure out how.

And last and maybe least: documentation.    There is none.  Watch out

Signed-off-by: utimatu <kylefriedline@gmail.com>